+++ packages/cli/create-medusa-app/src/utils/prepare-project.ts
@@ -2,20 +2,15 @@ import fs from "fs"
 import path from "path"
 import { Ora } from "ora"
 import execute from "./execute.js"
-import { EOL } from "os"
 import { displayFactBox, FactBoxOptions } from "./facts.js"
 import ProcessManager from "./process-manager.js"
 import { clearProject } from "./clear-project.js"
 import type { Client } from "pg"
+import getEnvVariables from "./get-env-variables.js"
 
 const ADMIN_EMAIL = "admin@medusa-test.com"
-// TODO remove preview links once we move to main docs
-export const STORE_CORS = "http://localhost:8000,https://docs.medusajs.com,https://medusa-docs-v2-git-docs-v2-medusajs.vercel.app,https://medusa-resources-git-docs-v2-medusajs.vercel.app"
-export const ADMIN_CORS = "http://localhost:7000,http://localhost:7001,https://docs.medusajs.com,https://medusa-docs-v2-git-docs-v2-medusajs.vercel.app,https://medusa-resources-git-docs-v2-medusajs.vercel.app"
-export const AUTH_CORS = ADMIN_CORS
-export const DEFAULT_REDIS_URL = "redis://localhost:6379"
 
-type PrepareOptions = {
+export type PrepareOptions = {
   directory: string
   dbConnectionString: string
   seed?: boolean
@@ -73,15 +68,12 @@ export default async ({
   let inviteToken: string | undefined = undefined
 
   // add environment variables
-  let env = `MEDUSA_ADMIN_ONBOARDING_TYPE=${onboardingType}${EOL}STORE_CORS=${STORE_CORS}${EOL}ADMIN_CORS=${ADMIN_CORS}${EOL}AUTH_CORS=${AUTH_CORS}${EOL}REDIS_URL=${DEFAULT_REDIS_URL}${EOL}JWT_SECRET=supersecret${EOL}COOKIE_SECRET=supersecret`
-
-  if (!skipDb) {
-    env += `${EOL}DATABASE_URL=${dbConnectionString}${EOL}POSTGRES_URL=${dbConnectionString}`
-  }
-
-  if (nextjsDirectory) {
-    env += `${EOL}MEDUSA_ADMIN_ONBOARDING_NEXTJS_DIRECTORY=${nextjsDirectory}`
-  }
+  let env = getEnvVariables({
+    onboardingType,
+    skipDb,
+    dbConnectionString,
+    nextjsDirectory
+  })
   
   fs.appendFileSync(path.join(directory, `.env`), env)
